#23 Refactor MainActivity
Refactor and cleaned the MainActivity.java class to have only what is actually used
