+++ EasyPlants_Complete/app/src/main/res/layout/activity_main.xml
@@ -18,26 +18,6 @@
         android:layout_marginBottom="112dp"
         android:text="See your plants here" />
 
-    <TextView
-        android:id="@+id/text_humidityText"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="132dp"
-        android:text="Welcome to"
-        android:textSize="24dp" />
-
-    <TextView
-        android:id="@+id/text_humidityReading"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="183dp"
-        android:text="EasyPlant"
-        android:textSize="24sp" />
-
     <com.gelitenight.waveview.library.WaveView
         android:id="@+id/wave"
         android:layout_width="300dp"
@@ -54,6 +34,15 @@
         android:layout_centerHorizontal="true"
         android:layout_marginBottom="37dp"
         android:text="Log Out" />
+
+    <ImageView
+        android:id="@+id/easy_plant_logo"
+        android:layout_width="wrap_content"
+        android:layout_height="188dp"
+        android:layout_alignParentStart="true"
+        android:layout_alignParentTop="true"
+        android:layout_marginTop="84dp"
+        app:srcCompat="@drawable/easy_plant_name" />
 
 
 </android.widget.RelativeLayout>
